Structure rebuild
Script scrutrure has been rebuilt
