+++ Diabolo/Assets/__Scripts/Control.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 60d39694840e7124f95fdcb443d94c38
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
