+++ Diabolo/Assets/__Scripts/Core/FollowCamera.cs
@@ -1,0 +1,21 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace RPG.Core
+{
+    public class FollowCamera : MonoBehaviour
+    {
+        [SerializeField] Transform target;
+
+        private void LateUpdate()
+        {
+            FollowPlayer();
+        }
+
+        private void FollowPlayer()
+        {
+            transform.position = target.position;
+        }
+    }
+}
