--- Diabolo/Assets/__Scripts/FollowCamera.cs
@@ -1,18 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class FollowCamera : MonoBehaviour
-{
-    [SerializeField] Transform target;
-
-    private void LateUpdate()
-    {
-        FollowPlayer();
-    }
-
-    private void FollowPlayer()
-    {
-        transform.position = target.position;
-    }
-}
