+++ Diabolo/Assets/__Scripts/Movement/Mover.cs
@@ -1,0 +1,39 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.AI;
+
+namespace RPG.Movement
+{
+    public class Mover : MonoBehaviour
+    {
+        [SerializeField] Transform target;
+
+        NavMeshAgent navMeshAgent;
+        Animator animator;
+
+        private void Awake()
+        {
+            navMeshAgent = GetComponent<NavMeshAgent>();
+            animator = GetComponent<Animator>();
+        }
+
+        private void Update()
+        {
+            UpdateAnimator();
+        }
+
+        public void MoveTo(Vector3 destination)
+        {
+            navMeshAgent.destination = destination;
+        }
+
+        private void UpdateAnimator()
+        {
+            Vector3 velocity = navMeshAgent.velocity;
+            Vector3 localVelocity = transform.InverseTransformDirection(velocity);
+            float speed = localVelocity.z;
+            animator.SetFloat("Vertical", speed);
+        }
+    }
+}
