--- Diabolo/Assets/__Scripts/Mover.cs
@@ -1,48 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.AI;
-
-public class Mover : MonoBehaviour
-{
-    [SerializeField] Transform target;
-
-    NavMeshAgent navMeshAgent;
-    Animator animator;
-
-    private void Awake()
-    {
-        navMeshAgent = GetComponent<NavMeshAgent>();
-        animator = GetComponent<Animator>();
-    }
-
-    private void Update()
-    {
-        if (Input.GetMouseButton(0))
-        {
-            MoveToTarget();
-        }
-
-        UpdateAnimator();
-    }
-
-    private void MoveToTarget()
-    {
-        Ray ray = Camera.main.ScreenPointToRay(Input.mousePosition);
-        RaycastHit hit;
-        bool hasHit = Physics.Raycast(ray, out hit);
-
-        if (hasHit)
-        {
-            navMeshAgent.destination = hit.point;
-        }
-    }
-
-    private void UpdateAnimator()
-    {
-        Vector3 velocity = navMeshAgent.velocity;
-        Vector3 localVelocity = transform.InverseTransformDirection(velocity);
-        float speed = localVelocity.z;
-        animator.SetFloat("Vertical", speed);
-    }
-}
